Remove UnsafeRun syntax, move unsafeRunSync and unsafeRunTimed to JVM only, added companion summoner
