--- kernel/shared/src/main/scala/cats/effect/unsafe/syntax/UnsafeRunSyntax.scala
@@ -1,42 +1,0 @@
-/*
- * Copyright 2020 Typelevel
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package cats.effect.unsafe.syntax
-
-import scala.concurrent.Future
-import scala.concurrent.duration.Duration
-
-import cats.effect.unsafe.UnsafeRun
-
-trait UnsafeRunSyntax {
-  implicit def unsafeRunOps[F[_], A](wrapped: F[A])(
-      implicit F: UnsafeRun[F]): UnsafeRunOps[F, A] =
-    new UnsafeRunOps(wrapped)
-}
-
-final class UnsafeRunOps[F[_], A](val wrapped: F[A])(implicit F: UnsafeRun[F]) {
-  def unsafeRunFutureCancelable: (Future[A], () => Future[Unit]) =
-    F.unsafeRunFutureCancelable(wrapped)
-
-  def unsafeRunAsync(): Unit =
-    F.unsafeRunAsync(wrapped)
-
-  def unsafeRunSync(): A =
-    F.unsafeRunSync(wrapped)
-
-  def unsafeRunTimed(timeout: Duration): A =
-    F.unsafeRunTimed(wrapped, timeout)
-}
